This change modifies the Connector Manager sources to compile cleanly under Java 5 (1.5).  Neary all Collections, Lists, Sets, Maps, etc, and several interfaces (like Comparable> were appropriately parameterized.
Particular attention was made to the SPI.

In general, I avoided easy-outs like <?> parameters and @SuppressWarnings,
however, they were often necessary when dealing with Spring getBean(),
JSON, and Properties.  [I have no idea why Properties.properyNames()
returns Enumeration <?>, rather than Enumeration <String>, or why Properties
implements Map <?, ?>, rather than Map <String, String>.]

I modified build.xml javac task as follows:
  - changed -target and -source switches from 1.4 to 1.5
  - additional warnings with "-Xlint -Xlint:-serial" which turns on most
    warnings, except serializaion UUID warnings.
  - javatest target aslo disables "deprecated" warnings to avoid flagging
    tests for deprecated features.

Going into this, my understanding of Generics was less than ideal,
so some things might have been done better.

There is probably some improvement that can be with parameterization
with Spring.

I replaced many uses of StringBuffer with the faster StringBuilder.

I replaced many Iterator based-loops with for-each loop syntax.

+++ changes/Brett.Michael.Johnson/java5_compile/trunk/projects/connector-manager/source/java/com/google/enterprise/connector/persist/GenerationalStateStore.java
@@ -128,7 +128,7 @@ public class GenerationalStateStore implements ConnectorStateStore {
       generation = new Long(currentGeneration(context));
       myGenerations.put(connectorName, generation);
     }
-    return generation.longValue();
+    return generation;
   }
 
   /**
@@ -140,7 +140,7 @@ public class GenerationalStateStore implements ConnectorStateStore {
   protected static long currentGeneration(StoreContext context) {
     synchronized (generations) {
       Long generation = generations.get(context.getConnectorName());
-      return (generation == null) ? 0 : generation.longValue();
+      return (generation == null) ? 0 : generation;
     }
   }
 
@@ -156,7 +156,7 @@ public class GenerationalStateStore implements ConnectorStateStore {
     synchronized (generations) {
       Long generation = generations.get(context.getConnectorName());
       generations.put(context.getConnectorName(),
-          new Long((generation == null) ? 1 : generation.longValue() + 1));
+          new Long((generation == null) ? 1 : generation + 1));
     }
   }
 }
