+++ changes/Brett.Michael.Johnson/java5_compile/trunk/projects/connector-manager/source/java/com/google/enterprise/connector/traversal/MimeTypeMap.java
@@ -164,7 +164,7 @@ public class MimeTypeMap {
         }
       }
     }
-    int sl = (result == null) ? unknownMimeTypeSupportLevel : result.intValue();
+    int sl = (result == null) ? unknownMimeTypeSupportLevel : result;
     if (LOGGER.isLoggable(Level.FINEST)) {
       LOGGER.finest("Mime type support level for " + mimeType + " is " + sl);
     }
